backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "kanban backend"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=6.0.4",
    "django-cors-headers>=4.9.0",
    "django-ninja>=1.6.2",
    "django-ninja-jwt>=5.3.0",
    "python-dotenv>=1.2.2",
    "requests>=2.33.1",
]
